feat: Omaha High Poker
The objective of this project is to design an algorithm and implement a C++ program that can determine the winning hand between two Omaha High Poker hands. The program will analyze the provided input file "distrib/input.txt" and generate the output file "distrib/output.txt" with the results of the hand comparison.